requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==5.0.3
django-coreapi==2.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
whitenoise==6.6.0
